added rtmidi header and tidied up csound struct further
