+++ include/coreDefs.h
@@ -153,6 +153,7 @@ static const uint32_t PHMASK = (1 << 24) - 1;
 
 #define NAMELEN 40              /* array size of repeat macro names */
 #define RPTDEPTH 40             /* size of repeat_n arrays (39 loop levels) */  
+#define LBUFSIZ   32768
   
   typedef int32_t (*SUBR)(CSOUND *, void *); 
   enum {FFT_LIB=0, PFFT_LIB, VDSP_LIB};
